+++ elasticache/elasticache.tf
@@ -22,6 +22,7 @@ resource "aws_elasticache_replication_group" "non_cluster_mode" {
   apply_immediately          = true
   multi_az_enabled           = true
   automatic_failover_enabled = true
+  data_tiering_enabled       = var.data_tiering_enabled
 
   security_group_ids = [
     "${aws_security_group.main.id}",