+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 This project does not follow SemVer, since modules are independent of each other; thus, SemVer does not make sense. Changes are grouped per module.
 
+## [unreleased]
+
+## ElastiCache
+- added `data_tiering_enabled` to the available options, defaulting to `false`
+
+
 ## [v2022.07.08]
 
 ## VPN